requirements.txt
pandas
tqdm
Pillow
tabula
numpy
pdf2image
pytesseract
langdetect
xlsxwriter
pywin32